pyproject.toml
[tool.poetry]
name = "grimoirelab-core"
version = "0.0.1"
description = "Core of GrimoireLab"
authors = [
    "GrimoireLab Developers"
]
license = "GPL-3.0+"

readme = "README.md"

homepage = "https://chaoss.github.io/grimoirelab/"
repository = "https://github.com/chaoss/grimoirelab-core"

keywords = [
    "development",
    "grimoirelab"
]

packages = [
    { include = "grimoirelab", from = "src" },
]

classifiers = [
   "Development Status :: 4 - Beta",
   "Intended Audience :: Developers",
   "Topic :: Software Development",
   "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
   "Programming Language :: Python :: 3"
]


[tool.poetry.urls]
"Bug Tracker" = "https://github.com/chaoss/grimoirelab-core/issues"

[tool.poetry.scripts]
grimoirelab = "grimoirelab.core.runner:grimoirelab"

[tool.poetry.dependencies]
python = "^3.8"

django = "^4.2"
django-rq = "^2.10.1"
mysqlclient = "2.0.3"
uWSGI = "^2.0"
grimoirelab-toolkit = {version = "^1.0.2", allow-prereleases = true}
perceval = {version = "^1.0.2", allow-prereleases = true}
# TODO: Change chronicler to package
grimoirelab-chronicler = {git = "https://github.com/grimoirelab/grimoirelab-chronicler.git", allow-prereleases = true}

[tool.poetry.dev-dependencies]
fakeredis = "^2.0.0"
httpretty = "^1.1.4"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
